Написана операция макспулинга, улучшен класс матрица, проверена работа.

+++ CNN/Filter.cpp
@@ -10,7 +10,7 @@ Filter::Filter(const int & i_, const int & j_ ) : Matrix(i_, j_)
 {
 }
 
-Filter::Filter( int ** arr_, const int & i_, const int & j_) : Matrix(arr_, i_, j_)
+Filter::Filter(double ** arr_, const int & i_, const int & j_) : Matrix(arr_, i_, j_)
 {
 }
